Do not catch `InvocationTargetException` in `getInstance`
Before, we caught the `InvocationTargetException` when `getInstance` was
called. This prevented the occurrence of the exception in the default
constructor. This patch removes the catch and adds throw annotations so that the
exception can be caught in the generated test method.

In the case of an exception in the static initializer, an
`ExceptionInInitializer` is thrown when doing `Class.forName(...)`. This error
needs to be caught and its cause is wrapped into the expected
`InvocationTargetException`.