Create constructor with same parameters as in superclass
Before, when an abstract class had no default constructor, our approach to add a
default one in the derived class worked because we caught the resulting
`MethodNotFoundException` thrown because the call to the super class constructor
did not work.

Now we simply create a constructor with the same types as one of the
constructors of the super class.

In the case of an interface, we add a default constructor.