Use `repo1.maven.org`
The above is a HTTP accessible mirror of maven central. According to

https://central.sonatype.org/articles/2018/May/04/discontinue-support-for-tlsv11-and-below/?__hstc=31049440.39bec4937410179f17181d98964c7446.1529416496779.1529416496779.1529416496779.1&__hssc=31049440.1.1529416496780&__hsfp=2831431618

it will be maintained in the long term.

The switch to HTTP is needed as openjdk-6 doesn't support TLS1.2 and Maven
switched to TLS1.2 access on June 18th 2018. An update to openjdk-6 is unlikely
as it has reached its EOL on Dec. 31st 2016,
cf. http://mail.openjdk.java.net/pipermail/jdk6-dev/2016-October/003606.html