Creates a new exception to wrap other exceptions in Reflector.
Reflector can throw a lot of exceptions, so therefore a test that uses Reflector can throw a lot of exceptions. We would like to list the specific exceptions a test method throws in the test, TG-4345, so this work squashes all of the exceptions Reflector can throw into a RuntimeException called DeeptestUtilsRuntimeException. InvocationTargetException is the only exception that has not been squash because deeptest relies on that being thrown at the right time.