Expose private overload of `setField`
This overload needs to be exposed outside of deeptest utils, so that it
can be used to set innaccessible, hidden fields. In this context a
hidden field is a field in a base class which has the same name as
another field which is defined in another class which extends the base
class.