Move exception wrapping down the call heirachy
The previously private overload of setField was declared to throw
various exceptions which now need to be wrapped. This commit removes the
declaration that this now public overload throws these various
exceptions by wrapping them inside this overload. Because the overload
which was previously doing the exception wrapping calls the other
overload, it no longer needs to do the wrapping.