Remove all references to real constructors
Calling real constructors is problematic as they might have side effects
that are not considered by Diffblue Cover.
For example, a field could be set by calling a method on a static field,
and Diffblue Cover would not know that this field should be set to
something non-null. See the following commit for an example.