Remove openjdk6 Travis build and pom profile
openjdk6 is no longer supported on Travis. Since we have source and
target for the compiler set to 1.6 in the pom file anyway, we just
remove the Java 6 build and corresponding profile for now. To make sure
we're testing compilation with other sources/targets/compilers, TG-5884
should be done in the future.