Move "make public" logic to where it is needed
We only need to declare methods, fields and constructors as public if we
expect to construct an implementing class / subclass that inherits them.
In the case where we are trying to instantiate a concrete class, we only
call Class.forName and ObjenesisStd.newInstance, both of which work for
(partly or fully) inaccessible classes.