Delete broken/untested code block
From what I can tell, this part of the code was supposed to go through
the abstract methods of an abstract class and for each of them, add a
trivial concrete method to the newly generated subclass.
There are two problems with it:
1) It doesn't always work. I tried a small example and found that it
works with public abstract methods, but it doesn't work with any other
access modifier than public. It also only works for _declared_ methods,
i.e. it does not work for any inherited methods.
2) Even if it did always work, it is not consistent with the behaviour
of test-generator. A trivial method would always return the default
value for its return type, but the caller might have different branches
based on the return type, which we would generate different tests for.
All but at most one of these tests would always fail. We should never
end up in this case anyway, because when we have a call to an abstract
method, we should mock it in the generated tests, rather than use a call
to Reflector.getInstance.

There are no tests for this code, and there generally seems to be no
reason to keep it.